responsive validate & wip refactorisation
