+++ src/main/java/com/Ecommerce/acme/controller/LandingController.java
@@ -11,6 +11,5 @@ public class LandingController {
 	public String landingPage(Model model) {
 		return "landingPage";
 	}
-	
 
 }
