+++ src/main/resources/templates/addCategory.html
@@ -1,19 +1,23 @@
 <!DOCTYPE html>
 
-<html lang="en" xmlns:th="http://www.thymeleaf.org"
-	xmlns:sec="https://www.thymeleaf.org/thymeleaf-extras-springsecurity5">
+<html lang="en" xmlns:th="http://www.thymeleaf.org" xmlns:sec="https://www.thymeleaf.org/thymeleaf-extras-springsecurity5">
 
 <head>
 <meta charset="utf-8">
 <title>Ajout Catégorie</title>
 <link th:href="@{/css/bootstrap.min.css}" rel="stylesheet" />
+<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.9.1/font/bootstrap-icons.css">
 </head>
 
-<body>
+<header>
 
-	<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
+<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
 		<div class="container-fluid">
-			<a class="navbar-brand" href="#">ACME Gestion Pro.</a>
+			<a class="navbar-brand" href="/admin/manage_category">ACME Gestion Pro.</a>
+			
+		<button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarColor01" aria-controls="navbarColor01" aria-expanded="false" aria-label="Toggle navigation">
+     			 <span class="navbar-toggler-icon"></span>
+   		</button>
 
 			<div class="collapse navbar-collapse" id="navbarColor01">
 
@@ -25,9 +29,9 @@
 					<li class="nav-item"><a class="nav-link"
 						href="/admin/manage_product">Collection</a></li>
 					<li class="nav-item"><a class="nav-link"
-						href="/admin/addAdmin">Profil-Admin</a></li>
-					<li class="nav-item"><a class="nav-link"
-						href="/admin/manage_user">Profil-Client</a></li>
+						href="/admin/addAdmin">Création-Admin</a></li>
+					<li class="nav-item"><a class="nav-link" href="/admin/manage_user">Profil-Client</a>
+					</li>
 					<li class="nav-item"><a class="nav-link active" href="/home">Interface-Client</a>
 					</li>
 				</ul>
@@ -40,6 +44,10 @@
 		</button>
 
 	</nav>
+
+</header>
+
+<body>
 
 	<div class="container">
 		<div class="row justify-content-center">
@@ -76,7 +84,7 @@
 			<p class="text-center">Ajout d'une Catégorie</p>
 		</div>
 
-		<div class="row justify-content-center">
+		<div class="row justify-content-center pb-5">
 			<div class="col-6">
 
 				<form th:action="@{/admin/add_category}" th:object="${category}"
@@ -101,4 +109,23 @@
 	</div>
 
 </body>
+
+	<script
+		src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/js/bootstrap.bundle.min.js"
+		integrity="sha384-ygbV9kiqUc6oa4msXn9868pTtWMgiQaeYH7/t7LECLbyPA2x65Kgf80OJFdroafW"
+		crossorigin="anonymous"></script>
+
+
+<footer class="bg-dark text-center text-white">
+	<div class="text-center p-3 navbar-dark bg-primary">
+		© 2022 Copyright: <a class="text-white" href="#">ACME.com</a>
+		<a href="https://github.com/sheerinbanu/acme" class="text-white" target="_blank">
+			<i class="bi bi-github"></i>
+		</a>
+		<a href="https://trello.com/b/y2dGQzit/acme" class="text-white" target="_blank">
+			<i class="bi bi-kanban"></i>
+		</a>
+	</div>
+</footer>
+
 </html>
