+++ src/main/resources/templates/admin.html
@@ -11,7 +11,7 @@
 
  	<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
   		<div class="container-fluid">
-   			 <a class="navbar-brand" href="#">ACME Gestion Pro Admin.</a>
+   			 <a class="navbar-brand" href="/admin/manage_category">ACME Gestion Pro Admin.</a>
   		</div>
  	</nav>
