+++ src/main/resources/templates/cart.html
@@ -7,13 +7,18 @@
 <meta charset="utf-8">
 <title>Panier</title>
 <link th:href="@{/css/bootstrap.min.css}" rel="stylesheet" />
+<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.9.1/font/bootstrap-icons.css">
 </head>
 
-<body>
+<header>
 
-	<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
+<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
 		<div class="container-fluid">
-			<a class="navbar-brand" href="#">ACME Gestion Pro.</a>
+			<a class="navbar-brand" href="/home">ACME Gestion Pro.</a>
+			
+			 <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarColor01" aria-controls="navbarColor01" aria-expanded="false" aria-label="Toggle navigation">
+     			 <span class="navbar-toggler-icon"></span>
+   			 </button>
 
 			<div class="collapse navbar-collapse" id="navbarColor01">
 
@@ -28,15 +33,24 @@
 					</li>
 					<li class="nav-item"><a class="nav-link" href="/cart">Panier</a>
 					</li>
+					
+					<li sec:authorize-url="/admin/" class="nav-item"> 
+						<a class="nav-link" id="admin" th:href="@{/admin/manage_category}">Interface-Admin</a>
+					</li>
+					
 				</ul>
-
 			</div>
 		</div>
 
 		<button class="btn btn-outline-danger my-2 my-sm-0">
 			<a onclick="document.forms['logoutForm'].submit()">Déconnexion</a>
 		</button>
+
 	</nav>
+
+</header>
+
+<body>
 
 	<div class="container">
 
@@ -182,9 +196,23 @@
 		</div>
 	</section>
 </body>
+
+<script
+		src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/js/bootstrap.bundle.min.js"
+		integrity="sha384-ygbV9kiqUc6oa4msXn9868pTtWMgiQaeYH7/t7LECLbyPA2x65Kgf80OJFdroafW"
+		crossorigin="anonymous"></script>
+
+
 <footer class="bg-dark text-center text-white">
 	<div class="text-center p-3 navbar-dark bg-primary">
 		© 2022 Copyright: <a class="text-white" href="#">ACME.com</a>
+		<a href="https://github.com/sheerinbanu/acme" class="text-white" target="_blank">
+			<i class="bi bi-github"></i>
+		</a>
+		<a href="https://trello.com/b/y2dGQzit/acme" class="text-white" target="_blank">
+			<i class="bi bi-kanban"></i>
+		</a>
 	</div>
 </footer>
+
 </html>
