+++ src/main/resources/templates/login.html
@@ -10,7 +10,7 @@
 
 	<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
 		<div class="container-fluid">
-			<a class="navbar-brand" href="#">ACME Gestion Pro.</a>
+			<a class="navbar-brand" href="/">ACME Gestion Pro.</a>
 		</div>
 	</nav>
