+++ src/main/resources/templates/manageProduct.html
@@ -7,14 +7,18 @@
 <meta charset="utf-8">
 <title>Gestion Produits</title>
 <link th:href="@{/css/bootstrap.min.css}" rel="stylesheet" />
+<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.9.1/font/bootstrap-icons.css">
 </head>
 
-<body>
+<header>
 
-
-	<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
+<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
 		<div class="container-fluid">
-			<a class="navbar-brand" href="#">ACME Gestion Pro.</a>
+			<a class="navbar-brand" href="/admin/manage_category">ACME Gestion Pro.</a>
+			
+		<button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarColor01" aria-controls="navbarColor01" aria-expanded="false" aria-label="Toggle navigation">
+     			 <span class="navbar-toggler-icon"></span>
+   		</button>
 
 			<div class="collapse navbar-collapse" id="navbarColor01">
 
@@ -26,7 +30,7 @@
 					<li class="nav-item"><a class="nav-link"
 						href="/admin/manage_product">Collection</a></li>
 					<li class="nav-item"><a class="nav-link"
-						href="/admin/addAdmin">Profil-Admin</a></li>
+						href="/admin/addAdmin">Création-Admin</a></li>
 					<li class="nav-item"><a class="nav-link" href="/admin/manage_user">Profil-Client</a>
 					</li>
 					<li class="nav-item"><a class="nav-link active" href="/home">Interface-Client</a>
@@ -41,6 +45,11 @@
 		</button>
 
 	</nav>
+
+</header>
+
+
+<body>
 
 	<div class="container">
 		<div class="row justify-content-center">
@@ -58,17 +67,12 @@
 	</div>
 
 	<div class="d-flex justify-content-around mt-5 mb-5 ">
-		<div class="table-responsive">
-			<form class="d-flex">
-				<input class="form-control me-sm-2" type="text"
-					placeholder="rechercher des produits">
-				<button class="btn btn-secondary my-2 my-sm-0" type="submit">rechercher</button>
-			</form>
-		</div>
+		
 		<div>
 			<a href="/admin/add_product" class="btn btn-outline-success">+
 				Ajouter un produit</a>
 		</div>
+		
 	</div>
 	
 	<div class="container">
@@ -109,4 +113,23 @@
 		</div>
 	</div>
 </body>
+
+	<script
+		src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/js/bootstrap.bundle.min.js"
+		integrity="sha384-ygbV9kiqUc6oa4msXn9868pTtWMgiQaeYH7/t7LECLbyPA2x65Kgf80OJFdroafW"
+		crossorigin="anonymous"></script>
+
+
+<footer class="bg-dark text-center text-white">
+	<div class="text-center p-3 navbar-dark bg-primary">
+		© 2022 Copyright: <a class="text-white" href="#">ACME.com</a>
+		<a href="https://github.com/sheerinbanu/acme" class="text-white" target="_blank">
+			<i class="bi bi-github"></i>
+		</a>
+		<a href="https://trello.com/b/y2dGQzit/acme" class="text-white" target="_blank">
+			<i class="bi bi-kanban"></i>
+		</a>
+	</div>
+</footer>
+
 </html>
