+++ src/main/resources/templates/manageUser.html
@@ -1,72 +1,111 @@
 <!DOCTYPE html>
-<html lang="fr" xmlns:th="http://www.thymeleaf.org">
+
+<html lang="en" xmlns:th="http://www.thymeleaf.org" xmlns:sec="https://www.thymeleaf.org/thymeleaf-extras-springsecurity5">
+
 <head>
 <meta charset="UTF-8">
+<title>Gestion des Commerciaux</title>
 <link th:href="@{/css/bootstrap.min.css}" rel="stylesheet" />
+<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.9.1/font/bootstrap-icons.css">
 </head>
+
 <header>
-	<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
+
+<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
 		<div class="container-fluid">
-			<div>
-				<img src="" th:src="@{/images/acmelogoD.svg}" width="10%"
-					height="10%" />
-			</div>
-			<button class="navbar-toggler" type="button"
-				data-bs-toggle="collapse" data-bs-target="#navbarColor01"
-				aria-controls="navbarColor01" aria-expanded="false"
-				aria-label="Toggle navigation">
-				<span class="navbar-toggler-icon"></span>
-			</button>
+			<a class="navbar-brand" href="/admin/manage_category">ACME Gestion Pro.</a>
+			
+		<button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarColor01" aria-controls="navbarColor01" aria-expanded="false" aria-label="Toggle navigation">
+     			 <span class="navbar-toggler-icon"></span>
+   		</button>
+
 			<div class="collapse navbar-collapse" id="navbarColor01">
+
 				<ul class="navbar-nav me-auto">
+					<li class="nav-item"><a class="nav-link" href="#">Commande</a>
+					</li>
 					<li class="nav-item"><a class="nav-link"
-						href="/admin/manage_order">Commandes </a></li>
+						href="/admin/manage_category">Categorie</a></li>
 					<li class="nav-item"><a class="nav-link"
 						href="/admin/manage_product">Collection</a></li>
-					<li class="nav-item"><a class="nav-link active"
-						href="/admin/manage_user">Profils</a> <span
-						class="visually-hidden">(current)</span></li>
 					<li class="nav-item"><a class="nav-link"
-						href="/admin/manage_category">Categories</a>
-					<li class="nav-item"><a href="#"
-						class="btn btn-outline-danger">Deconnexion</a></li>
-
+						href="/admin/addAdmin">Création-Admin</a></li>
+					<li class="nav-item"><a class="nav-link" href="/admin/manage_user">Profil-Client</a>
+					</li>
+					<li class="nav-item"><a class="nav-link active" href="/home">Interface-Client</a>
+					</li>
 				</ul>
 
 			</div>
 		</div>
+
+		<button class="btn btn-outline-danger my-2 my-sm-0">
+			<a onclick="document.forms['logoutForm'].submit()">Déconnexion</a>
+		</button>
+
 	</nav>
+
 </header>
+
 <body>
-	<div class="btn-group" role="group" aria-label="Basic example"></div>
+
 	<div class="container">
-		<h3>Tableau de bord</h3>
+		<div class="row justify-content-center">
+			<div class="col-6 mt-3" sec:authorize="isAuthenticated()">
+				<form id="logoutForm" method="POST" th:action="@{/logout}">
+					<input type="hidden" name="${_csrf.parameterName}"
+						value="${_csrf.token}" />
+				</form>
+					<h2>
+						Accès Portail <span sec:authentication="authorities[0]"></span>
+						ACME Pro.
+					</h2>
+			</div>
+		</div>
 	</div>
+	
+	<div class="d-flex justify-content-around mt-5 mb-5 ">
+		
+		<div>
+			<a href="/registration" class="btn btn-outline-success">+
+				Ajouter un commercial</a>
+		</div>
+		
+	</div>
+
 	<div class="container">
+		
 		<div class="table-responsive">
 			<table class="table table-hover">
 
 				<tr class="table-dark">
+				
 					<th><input class="form-check-input" type="checkbox" value=""
 						id="flexCheckDefault"> <label class="form-check-label"
 						for="flexDefault"></label></th>
 
-					<th scope="row">Numero commercial</th>
-					<th>nom</th>
-					<th>prenom</th>
-					<th>email</th>
-					<th>editer</th>
+					<th scope="row">Id</th>
+					<th>Nom</th>
+					<th>Prenom</th>
+					<th>Email</th>
+					<th>Marge</th>
+					<th>Username</th>
+					<th>Editer</th>
 					<th>Supprimer</th>
 
 				</tr>
 				<tr th:each="user : ${users}">
+				
 					<th><input class="form-check-input" type="checkbox" value=""
 						id="flexCheckDefault"><label class="form-check-label"
 						for="flexCheckDefault"> </label></th>
+						
 					<td th:text="${user.id_user}">...</td>
 					<td th:text="${user.last_name}">...</td>
 					<td th:text="${user.first_name}">...</td>
 					<td th:text="${user.email}">...</td>
+					<td th:text="${user.margin_rate}">...</td>
+					<td th:text="${user.username}">...</td>
 					<td><a th:href="@{'/admin/update_user/'+${user.id_user}}"
 						role="button" class="btn btn-success">&#9999</a></td>
 					<td><a th:href="@{'/admin/delete_user/' + ${user.id_user}}"
@@ -75,8 +114,26 @@
 			</table>
 		</div>
 	</div>
-	<div>
-		<a th:href="@{'/registration'}" role="button"
-			class="btn btn-outline-success position-end">Ajouter un
-			utilisateur</a>
+
+	
+</body>
+
+	<script
+		src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/js/bootstrap.bundle.min.js"
+		integrity="sha384-ygbV9kiqUc6oa4msXn9868pTtWMgiQaeYH7/t7LECLbyPA2x65Kgf80OJFdroafW"
+		crossorigin="anonymous"></script>
+
+
+<footer class="bg-dark text-center text-white">
+	<div class="text-center p-3 navbar-dark bg-primary">
+		© 2022 Copyright: <a class="text-white" href="#">ACME.com</a>
+		<a href="https://github.com/sheerinbanu/acme" class="text-white" target="_blank">
+			<i class="bi bi-github"></i>
+		</a>
+		<a href="https://trello.com/b/y2dGQzit/acme" class="text-white" target="_blank">
+			<i class="bi bi-kanban"></i>
+		</a>
 	</div>
+</footer>
+
+</html>
