+++ src/main/resources/templates/order.html
@@ -6,12 +6,18 @@
 <meta charset="UTF-8">
 <title>Produits</title>
 <link th:href="@{/css/bootstrap.min.css}" rel="stylesheet" />
+<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.9.1/font/bootstrap-icons.css">
 </head>
 
-<body>
-	<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
+<header>
+
+<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
 		<div class="container-fluid">
-			<a class="navbar-brand" href="#">ACME Gestion Pro.</a>
+			<a class="navbar-brand" href="/home">ACME Gestion Pro.</a>
+			
+			 <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarColor01" aria-controls="navbarColor01" aria-expanded="false" aria-label="Toggle navigation">
+     			 <span class="navbar-toggler-icon"></span>
+   			 </button>
 
 			<div class="collapse navbar-collapse" id="navbarColor01">
 
@@ -26,6 +32,11 @@
 					</li>
 					<li class="nav-item"><a class="nav-link" href="/cart">Panier</a>
 					</li>
+					
+					<li sec:authorize-url="/admin/" class="nav-item"> 
+						<a class="nav-link" id="admin" th:href="@{/admin/manage_category}">Interface-Admin</a>
+					</li>
+					
 				</ul>
 			</div>
 		</div>
@@ -35,6 +46,10 @@
 		</button>
 
 	</nav>
+
+</header>
+
+<body>
 
 	<div class="container">
 
@@ -62,7 +77,7 @@
 
 			<div class="row">
 
-				<div class="col-md-8 mx-auto">
+				<div class="col-md-12 mx-0">
 
 					<table
 						class="table bg-white rounded border table-striped table-dark">
@@ -86,7 +101,7 @@
 
 
 								<td>
-									<button type="button" class="btn btn-primary"  
+									<button type="button" class="btn btn-primary" th:data-name="${order.id_order}" th:onclick="getNumberCart(this.getAttribute('data-name'))" 
 										data-bs-toggle="modal" data-bs-target="#exampleModal">Détail</button>
 
 									<div class="modal fade" id="exampleModal" tabindex="-1"
@@ -97,6 +112,8 @@
 												<div class="modal-header">
 													<h5 class="modal-title" id="exampleModalLabel">Détails
 														de la Commande</h5>
+														
+													<input type="text" id="my_field" th:value="0" th:name="my_field">
 
 													<button type="button" class="btn-close"
 														data-bs-dismiss="modal" aria-label="Close"></button>
@@ -150,15 +167,29 @@ s
 		</div>
 	</section>
 
-	<script
-		src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/js/bootstrap.bundle.min.js"
+	<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/js/bootstrap.bundle.min.js"
 		integrity="sha384-ygbV9kiqUc6oa4msXn9868pTtWMgiQaeYH7/t7LECLbyPA2x65Kgf80OJFdroafW"
 		crossorigin="anonymous"></script>
+		
+	<script th:inline="javascript">
+			function getNumberCart(num) {		
+				alert("number order: " + num);
+				document.getElementById("my_field").value = num;				
+   			 }
+	</script>	
 
 </body>
+
 <footer class="bg-dark text-center text-white">
 	<div class="text-center p-3 navbar-dark bg-primary">
 		© 2022 Copyright: <a class="text-white" href="#">ACME.com</a>
+		<a href="https://github.com/sheerinbanu/acme" class="text-white" target="_blank">
+			<i class="bi bi-github"></i>
+		</a>
+		<a href="https://trello.com/b/y2dGQzit/acme" class="text-white" target="_blank">
+			<i class="bi bi-kanban"></i>
+		</a>
 	</div>
 </footer>
+
 </html>
